bugfix v2

+++ src/main/java/com/hongwei/model/v1/covid19/vic/VicDataSetsSource.kt
@@ -3,11 +3,11 @@ package com.hongwei.model.v1.covid19.vic
 data class VicDataSetsSource(
         val postcode: Int,
         val population: Long = 0L,
-        val active: Int,
-        val cases: Int,
+        val active: Long,
+        val cases: Long,
         val rate: Float,
-        val new: Int,
-        val band: Int,
+        val new: Long,
+        val band: Long,
         val data_date: String,
         val file_processed_date: String
 )
